Load tournament data from file
Also added FA Cup season 2022-23 data for testing

+++ src/main/java/org/openengine/pureengine/tournament/Tournament.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public class Tournament {
 
-    private long year;
+    private int year;
     private List<Team> participants;
     private Competition competition;
     private List<TournamentRound> rounds = new ArrayList<>();
@@ -19,7 +19,7 @@ public class Tournament {
         this(tournament.year, tournament.participants, tournament.competition);
     }
 
-    public Tournament(long year, List<Team> participants, Competition competition) {
+    public Tournament(int year, List<Team> participants, Competition competition) {
         this.year = year;
         this.participants = participants;
         this.competition = competition;
